+++ app/controllers/folio/console/pages_controller.rb
@@ -7,9 +7,9 @@ class Folio::Console::PagesController < Folio::Console::BaseController
     @catalogue_options = {}
 
     if Rails.application.config.folio_pages_ancestry
-      @catalogue_model = Folio::Page.filter_by_params(filter_params)
-                                    .arrange(order: :position)
+      @pages = @pages.accessible_by(current_ability, self.class.cancancan_accessible_by_action)
 
+      @catalogue_model = @pages.arrange(order: :position)
       @catalogue_options = { ancestry: true }
     else
       super